Moved Windows files into Windows directory from dll directory
