+++ src/Windows/dcap_provider_tests/dcap_provider_tests.vcxproj
@@ -156,13 +156,13 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\private.h" />
-    <ClInclude Include="..\dll\curl_easy.h" />
+    <ClInclude Include="..\curl_easy.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\logging.cpp" />
-    <ClCompile Include="..\dll\curl_easy.cpp" />
+    <ClCompile Include="..\curl_easy.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
