+++ src/Windows/local_cache.cpp
